# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DEV-001 Thyroid Pre-Visit Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create appointment invitation (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentCreated"
        "409":
          description: Possible duplicate patient detected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentCreated"

    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by date (YYYY-MM-DD)
        - name: doctorId
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by doctor ID
        - name: status
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by status
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment detail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Appointment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Edit appointment (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}/cancel:
    post:
      operationId: cancelAppointment
      tags: [appointments]
      summary: Cancel appointment (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /appointments/{id}/resend-link:
    post:
      operationId: resendLink
      tags: [appointments]
      summary: Resend magic link to patient (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Link resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResendLinkResult"
  /appointments/{id}/reopen:
    post:
      operationId: reopenAppointment
      tags: [appointments]
      summary: Reopen locked questionnaire (staff)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reopened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── Questionnaires ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    patientAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth ────────────────────────────────────────────────────────────────
...
    StaffUser:
      type: object
      required: [id, email, role, fullName, isActive, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [doctor, clinic_admin, nurse]
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        mfaEnabled:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
...
    Appointment:
      type: object
      required: [id, invitedFullName, invitedPhone, appointmentType, scheduledAt, status, createdAt]
      properties:
        id:
          type: string
        patientId:
          type: ["string", "null"]
        doctorId:
          type: string
        invitedFullName:
          type: string
        invitedPhone:
          type: string
        appointmentType:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft_invitation, link_sent, opened, in_progress, submitted, locked, reopened, rescheduled, cancelled]
        labStatus:
          type: ["string", "null"]
        excludedFromClinicalViews:
          type: boolean
        originalScheduledAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        doctor:
          $ref: "#/components/schemas/StaffUser"
        patient:
          $ref: "#/components/schemas/Patient"
    AppointmentDetail:
      type: object
      required: [id, invitedFullName, invitedPhone, appointmentType, scheduledAt, status, createdAt]
      properties:
        id:
          type: string
        patientId:
          type: ["string", "null"]
        doctorId:
          type: string
        invitedFullName:
          type: string
        invitedPhone:
          type: string
        appointmentType:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
        labStatus:
          type: ["string", "null"]
        excludedFromClinicalViews:
          type: boolean
        originalScheduledAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        doctor:
          $ref: "#/components/schemas/StaffUser"
        patient:
          $ref: "#/components/schemas/Patient"
        questionnaire:
          $ref: "#/components/schemas/QuestionnaireRecord"
        summaries:
          type: array
          items:
            $ref: "#/components/schemas/Summary"
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
    AppointmentInput:
      type: object
      required: [invitedFullName, invitedPhone, dateOfBirth, doctorId, appointmentType, scheduledAt]
      properties:
        invitedFullName:
          type: string
        invitedPhone:
          type: string
        dateOfBirth:
          type: string
          description: "YYYY-MM-DD — used for patient identity verification"
        doctorId:
          type: string
        appointmentType:
          type: string
        scheduledAt:
          type: string
          format: date-time
    AppointmentUpdate:
      type: object
      properties:
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        appointmentType:
          type: ["string", "null"]
        invitedPhone:
          type: ["string", "null"]
        doctorId:
          type: ["string", "null"]
    AppointmentCreated:
      type: object
      required: [appointment]
      properties:
        appointment:
          $ref: "#/components/schemas/Appointment"
        link:
          type: object
          properties:
            token:
              type: string
            url:
              type: string
        possibleDuplicate:
          type: boolean
    ResendLinkResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        linkUrl:
          type: ["string", "null"]

    # ── Patients ────────────────────────────────────────────────────────────
    Patient:
      type: object
      required: [id, fullName, phone, dateOfBirth, createdAt]
      properties:
        id:
          type: string
        fullName:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        sex:
          type: ["string", "null"]
        heightCm:
          type: ["string", "null"]
        matchStatus:
          type: string
        duplicateReviewFlag:
          type: boolean
        consentGivenAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    QuestionnaireRecord:
      type: object
      required: [id, appointmentId, status, schemaVersion, createdAt]
      properties:
        id:
          type: string
        appointmentId:
          type: string
        status:
          type: string
          enum: [not_started, consent_given, in_progress, saved, submitted, locked]
        schemaVersion:
          type: string
        answers:
          type: object
          additionalProperties: true
        consentGivenAt:
          type: ["string", "null"]
          format: date-time
        savedAt:
          type: ["string", "null"]
          format: date-time
        submittedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Document:
      type: object
      required: [id, appointmentId, originalFileName, mimeType, fileSizeBytes, createdAt]
      properties:
        id:
          type: string
        appointmentId:
          type: string
        originalFileName:
          type: string
        mimeType:
          type: string
        fileSizeBytes:
          type: integer
        documentType:
          type: ["string", "null"]
        labStatus:
          type: ["string", "null"]
        deletedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Summary:
      type: object
      required: [id, appointmentId, variant, content, generatedAt, templateVersion]
      properties:
        id:
          type: string
        appointmentId:
          type: string
        variant:
          type: string
          enum: [current_visit, current_visit_plus_history]
        content:
          type: string
        generatedAt:
          type: string
          format: date-time
        templateVersion:
          type: string
...